# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "K8sVision 后端 API 文档",
        "title": "K8sVision API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/ingress": {
            "get": {
                "description": "支持分页\n获取指定命名空间下的 Ingress 详情",
                "tags": [
                    "Ingress",
                    "Ingress"
                ],
                "summary": "获取 Ingress 详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "偏移量",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.APIResponse"
                        }
                    }
                }
            }
        },
        "/ingress/{namespace}/{name}": {
            "get": {
                "description": "支持分页\n获取指定命名空间下的 Ingress 详情",
                "tags": [
                    "Ingress",
                    "Ingress"
                ],
                "summary": "获取 Ingress 详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "偏移量",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "命名空间",
                        "name": "namespace",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Ingress 名称",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.APIResponse"
                        }
                    }
                }
            }
        },
        "/jobs": {
            "get": {
                "description": "支持分页\n获取指定命名空间下的 Job 详情",
                "tags": [
                    "Job",
                    "Job"
                ],
                "summary": "获取 Job 详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "偏移量",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.APIResponse"
                        }
                    }
                }
            }
        },
        "/jobs/{namespace}/{name}": {
            "get": {
                "description": "支持分页\n获取指定命名空间下的 Job 详情",
                "tags": [
                    "Job",
                    "Job"
                ],
                "summary": "获取 Job 详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "偏移量",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "命名空间",
                        "name": "namespace",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Job 名称",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.APIResponse"
                        }
                    }
                }
            }
        },
        "/namespaces": {
            "get": {
                "description": "获取所有命名空间\n获取指定命名空间详情",
                "tags": [
                    "Namespace",
                    "Namespace"
                ],
                "summary": "获取 Namespace 详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "命名空间",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.APIResponse"
                        }
                    }
                }
            }
        },
        "/namespaces/{name}": {
            "get": {
                "description": "获取所有命名空间\n获取指定命名空间详情",
                "tags": [
                    "Namespace",
                    "Namespace"
                ],
                "summary": "获取 Namespace 详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "命名空间",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/nodes/{name}": {
            "get": {
                "description": "获取集群节点列表\n获取指定节点详情",
                "tags": [
                    "Node",
                    "Node"
                ],
                "summary": "获取 Node 详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Node 名称",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.APIResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "page": {
                    "$ref": "#/definitions/model.PageMeta"
                },
                "timestamp": {
                    "type": "integer"
                },
                "traceId": {
                    "type": "string"
                }
            }
        },
        "model.PageMeta": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        }
    }
}